Player animations in and black lines on tile map fixed
Animations for the snake segments aren't implemented and it seems like a hassle... it'll be on the backlog. Next  priority is going to be implementing a ongoing counter for the number of apples the player collects. After that I'm thinking of adding some free unity sound effects and maybe a music loop and calling it a day for this